fix bug

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode6032/hardware/DuckSpinner.java
@@ -17,6 +17,6 @@ public class DuckSpinner {
     }
 
     public void setEnabled(int dir) {
-        spinner.setPower(dir * SPEED);
+        spinner.setPower(((float)dir) * SPEED);
     }
 }
